+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Features
+
+- If `footer` constructor argument is `null`, set empty Map
+
 ## [1.1.0] - 20220-10-25
 
 ### Features
@@ -14,5 +20,6 @@
 - Conventional commit parser
 - Conventional commit generator
 
+[Unreleased]: https://github.com/clean-code-rocks/conventional-commit/compare/v1.1.0...main
 [1.1.0]: https://github.com/clean-code-rocks/conventional-commit/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/clean-code-rocks/conventional-commit/releases/tag/v1.0.0